move mock data files to orgs folder
