+++ e2e/mockData/orgs/KPD/index.ts
@@ -1,4 +1,4 @@
-import { ZetkinOrganization } from '../../../src/types/zetkin';
+import { ZetkinOrganization } from '../../../../src/types/zetkin';
 
 const KPD: ZetkinOrganization = {
     id: 1,
